Pywb-style CDXJ input and output formats
This adds support for Pywb's CDXJ format. We follow the Pywb convention
of emitting numeric values as JSON strings but accept JSON numbers if
given them as input.

Support for arbitrarily named extension fields is not included yet and
will be added separately as it requires a new version of the index
storage format. Similarly our current index version doesn't really
support the notion of missing fields so we map missing fields to "-" or
-1 as appropriate for storage, which is a bit hacky but should generally
work for now.

Other proposed CDXJ variants (such as "OpenWayback CDXJ") are not
supported.